Map from SleParser to common format
The output of the SleParsers is a flattened list of dictionaries, which each dict containing the data and metadata of one spectrum (including the individual channel scans).

This list is mapped to a dictionary that has a similar structure as the output of the XmlSpecs parser, which can then be used to write the nxmpes structure using config files.