Reader for native SPECS SLE format (#110)
* Add reader for .sle XPS files (SPECS Prodigy)

Initial commit for XPS reader for SLE files (native format of SPECS Prodigy)
- Add sle_parsers which read the schedule (as xml) as well as the data from the SLE datafiles (which are SQLlite databases)
 - Implement interface in dataconverter/xps/reader_utils for SLE readers for different versions of .sle files

Currently missing:
 - mapping from output format of parse_file method of the different SleParsers to get_dict method of parse_xml of XML reader -> SleParser output cannot be mapped to Nxmpes at the moment

* Pylinting, minor naming updates

* Map from SleParser to common format

The output of the SleParsers is a flattened list of dictionaries, which each dict containing the data and metadata of one spectrum (including the individual channel scans).

This list is mapped to a dictionary that has a similar structure as the output of the XmlSpecs parser, which can then be used to write the nxmpes structure using config files.

* Small updates to SleParsers

 - Individual channel scans are stored in the data field of the spectrum dictionaries
- Fix some data type errors during reading
- Read out transmission data.

Note: Channel readout and calibration is currently not working ( see method _get_calibrated_data). Therefore, the "calibrated" data is currently just the scan of the first channel.

* Small changes to reader.py, multiple config files

- Restructure CONVERT_DICT
- Add "NXmpes_xps" to supported NXDLs
- Config files: either the config files in the xps folder are used (based on the file extension of the input) or a config file can be explicitly passed during the reading

* Fix error in transmission data parsing

- Read both TF data and the file it was stored in.

* Allow kwargs passing to the SleParsers

* Small update to config file parsing

* Pycodestyle linting

* Temporarily remove NXmpes_xps from supported_nxdls

Since there is no NXDL for the NXmpes_xps  application definition, the pytests are failing. Temporarily remove it until NXDL is available.

* Implement abstractmethod for flattening XML

Different SLE parsers have different xml schedules. Implement abstractmethod for abstract SleParser base class.

* pylinting

* Linting

* some more linting

* Mypy changes